Refactor: LadderData -> LadderDataInput으로 이름 변경 및 ui 레이어로 이동
