+++ src/main/java/ladder/domain/Ladder.java
@@ -3,6 +3,8 @@ package ladder.domain;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import ladder.ui.LadderDataInput;
+
 public class Ladder {
 
     private final Player player;
@@ -10,7 +12,7 @@ public class Ladder {
     private final LadderLine ladderLine;
     private final LadderResult ladderResult;
 
-    public Ladder(LadderData ladderData) {
+    public Ladder(LadderDataInput ladderData) {
         this(ladderData.getPlayer(), ladderData.getReward(), ladderData.getHeight());
     }
